Fixed bug in type guessing.
This fixes a bug where _guess_types() did not always determine numerical types correctly.  If _guess_types() examined a column of mixed floating-point and integer numbers, and the first value in the column happened to be an integer, then the code would always conclude that it was an integer column, which is obviously incorrect.  If the first value was a floating point, it worked as expected.  This was a consequence of the math that was used in line 75 of the script, and the check later (line 82) to see if the final values are equal to 1.  It should now work correctly for all cases.